requirements.txt
streamlit
transformers
torch
nltk
scikit-learn
pandas
numpy
matplotlib
newspaper3k
